+++ lib/girl_friday/work_queue.rb
@@ -19,6 +19,7 @@ module GirlFriday
       @total_processed = @total_errors = @total_queued = 0
       @persister = (options[:store] || Store::InMemory).new(name, (options[:store_config] || []))
       start
+      GirlFriday.add_queue WeakRef.new(self)
     end
 
     def self.immediate!